move preproc to funcs
