+++ yucca/functional/visualization/__init__.py
@@ -1,0 +1,1 @@
+from yucca.functional.visualization.imshow import get_train_fig_with_inp_out_tar
